requirements.txt
numpy==2.0.2
matplotlib==3.9.4
pillow==11.1.0
